third_party/blink/web_tests/platform/generic/fast/css/shorthand-setProperty-important-expected.txt

Verify that the priority parameter is a case-insensitive match to 'important' or the empty string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.style.getPropertyValue('border-bottom-style') is ""
PASS e.style.getPropertyPriority('border-bottom-style') is ''
PASS e.style.getPropertyValue('border-bottom-style') is ""
PASS e.style.getPropertyPriority('border-bottom-style') is ''
PASS e.style.getPropertyValue('border-bottom-style') is 'solid'
PASS e.style.getPropertyPriority('border-bottom-style') is 'important'
PASS e.style.getPropertyValue('border-bottom-style') is ""
PASS e.style.getPropertyPriority('border-bottom-style') is ''
PASS e.style.getPropertyValue('border-bottom-style') is ""
PASS e.style.getPropertyPriority('border-bottom-style') is ''
PASS e.style.getPropertyValue('border-bottom-style') is 'solid'
PASS e.style.getPropertyPriority('border-bottom-style') is ''
PASS successfullyParsed is true

TEST COMPLETE
